requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
Box2D==2.3.2
Box2D-kengz==2.3.3
cachetools==5.3.2
casadi==3.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.1.1
csnlp==1.5.8
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
idna==3.4
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpcrl==1.1.8
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==4.25.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
swig==4.1.1
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
torch==2.1.0
traitlets==5.13.0
triton==2.1.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.14.1
